Change internal behavior of Service Accounts
Instead of using properties for what we need specifically, we store
the full provided data, because we will pass it to the Google Cloud
Core component for authentication.

All we really need for our purposes is the project ID, and only from
within the factory.

That's why we also declare the ServiceAccount class internal from
now on.