Add `startAfter` and `endBefore` RTDB query filters
At the moment they don't seem to have an effect on the returned results
(just as if they didn't exist); it's unclear if the implementation is
incorrect or if the REST API doesn't support the new filters yet. If
you see why it's not working or if it _does_ work for you, please
let me know.